# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Haven Backend HTTP API
  version: 0.1.0
  description: |
    OpenAPI document for the Haven backend HTTP surface.

    Base path is `/api/v1`.
    WebSocket upgrade endpoints (`/ws` and `/realtime/ws`) are not represented here.
    Current WebSocket auth uses an initial `authenticate` message with a bearer token after connect.
    Query parameter token auth for WebSocket connections is no longer supported.
servers:
  - url: http://localhost:8086/api/v1
    description: Local development
...
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Register user
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: Registered user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserResponse'
        '400':
          $ref: '#/components/responses/ErrorBadRequest'
        '409':
          $ref: '#/components/responses/ErrorConflict'
        '422':
          $ref: '#/components/responses/ErrorValidation'
        '429':
          $ref: '#/components/responses/ErrorTooManyRequests'
        '500':
          $ref: '#/components/responses/ErrorInternal'
  /auth/login/challenge:
    post:
      tags: [Auth]
      summary: Start SRP login
      operationId: loginChallenge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginChallengeRequest'
      responses:
        '200':
          description: SRP challenge payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginChallengeResponse'
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
        '403':
          $ref: '#/components/responses/ErrorForbidden'
        '422':
          $ref: '#/components/responses/ErrorValidation'
        '429':
          $ref: '#/components/responses/ErrorTooManyRequests'
        '500':
          $ref: '#/components/responses/ErrorInternal'
  /auth/login/verify:
    post:
      tags: [Auth]
      summary: Complete SRP login
      operationId: loginVerify
      parameters:
        - name: x-srp-challenge-id
          in: header
          required: true
          description: Challenge id returned by `/auth/login/challenge`.
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginVerifyRequest'
      responses:
        '200':
          description: Verified session and issued tokens.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginVerifyResponse'
        '400':
          $ref: '#/components/responses/ErrorBadRequest'
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
        '403':
          $ref: '#/components/responses/ErrorForbidden'
        '422':
          $ref: '#/components/responses/ErrorValidation'
        '429':
          $ref: '#/components/responses/ErrorTooManyRequests'
        '500':
          $ref: '#/components/responses/ErrorInternal'
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh tokens
      operationId: refreshTokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Refreshed access and refresh tokens.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokens'
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
        '403':
          $ref: '#/components/responses/ErrorForbidden'
        '422':
          $ref: '#/components/responses/ErrorValidation'
        '429':
          $ref: '#/components/responses/ErrorTooManyRequests'
        '500':
          $ref: '#/components/responses/ErrorInternal'
  /auth/me:
    get:
      tags: [Auth]
      summary: Current authenticated user
      operationId: getAuthMe
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserResponse'
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
        '403':
          $ref: '#/components/responses/ErrorForbidden'
        '429':
          $ref: '#/components/responses/ErrorTooManyRequests'
        '500':
          $ref: '#/components/responses/ErrorInternal'
  /auth/email/verification/request:
    post:
      tags: [Auth]
      summary: Request email verification code
      operationId: requestEmailVerification
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailVerificationRequest'
      responses:
        '200':
          description: Verification request accepted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
        '403':
          $ref: '#/components/responses/ErrorForbidden'
        '422':
          $ref: '#/components/responses/ErrorValidation'
        '429':
          $ref: '#/components/responses/ErrorTooManyRequests'
        '500':
          $ref: '#/components/responses/ErrorInternal'
  /auth/email/verification/confirm:
    post:
      tags: [Auth]
      summary: Confirm email verification code
      operationId: confirmEmailVerification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailVerificationConfirmRequest'
      responses:
        '200':
          description: Email confirmed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '400':
          $ref: '#/components/responses/ErrorBadRequest'
        '422':
          $ref: '#/components/responses/ErrorValidation'
        '429':
          $ref: '#/components/responses/ErrorTooManyRequests'
        '500':
          $ref: '#/components/responses/ErrorInternal'
  /auth/2fa/setup:
    post:
      tags: [Auth]
      summary: Start two-factor setup
      operationId: setupTwoFactor
      security:
        - bearerAuth: []
      responses:
        '200':
          description: TOTP setup payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TwoFactorSetupResponse'
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
        '403':
          $ref: '#/components/responses/ErrorForbidden'
        '429':
          $ref: '#/components/responses/ErrorTooManyRequests'
        '500':
          $ref: '#/components/responses/ErrorInternal'
  /auth/2fa/confirm:
    post:
      tags: [Auth]
      summary: Confirm two-factor setup
      operationId: confirmTwoFactor
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TwoFactorConfirmRequest'
      responses:
        '200':
          description: Two-factor confirmed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
        '403':
          $ref: '#/components/responses/ErrorForbidden'
        '422':
          $ref: '#/components/responses/ErrorValidation'
        '429':
          $ref: '#/components/responses/ErrorTooManyRequests'
        '500':
          $ref: '#/components/responses/ErrorInternal'
  /auth/2fa/disable:
    post:
      tags: [Auth]
      summary: Disable two-factor
      operationId: disableTwoFactor
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TwoFactorDisableRequest'
      responses:
        '200':
          description: Two-factor disabled.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
        '403':
          $ref: '#/components/responses/ErrorForbidden'
        '422':
          $ref: '#/components/responses/ErrorValidation'
        '429':
          $ref: '#/components/responses/ErrorTooManyRequests'
        '500':
          $ref: '#/components/responses/ErrorInternal'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: PASETO
...
  responses:
    ErrorBadRequest:
      description: Bad request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
    ErrorUnauthorized:
      description: Unauthorized.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
    ErrorForbidden:
      description: Forbidden.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
...
    ErrorConflict:
      description: Conflict.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
    ErrorValidation:
      description: Validation error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
    ErrorTooManyRequests:
      description: Rate limit hit.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
...
    ErrorInternal:
      description: Internal server error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
  schemas:
    ErrorBody:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    AuthTokens:
      type: object
      required: [access_token, refresh_token, token_type, expires_in_seconds]
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          example: Bearer
        expires_in_seconds:
          type: integer
          format: int64
    RegisterRequest:
      type: object
      required:
        - username
        - display_name
        - email
        - srp_salt
        - srp_verifier
        - date_of_birth
        - locale
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 32
        display_name:
          type: string
          minLength: 1
          maxLength: 64
        email:
          type: string
          format: email
        srp_salt:
          type: string
          description: Base64 encoded SRP salt from client.
        srp_verifier:
          type: string
          description: Base64 encoded SRP verifier from client.
        date_of_birth:
          type: string
          format: date
        locale:
          type: string
          minLength: 2
          maxLength: 10
    LoginChallengeRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
    LoginChallengeResponse:
      type: object
      required: [challenge_id, srp_salt, server_public_key_b]
      properties:
        challenge_id:
          type: string
        srp_salt:
          type: string
        server_public_key_b:
          type: string
    LoginVerifyRequest:
      type: object
      required: [email, client_public_key_a, client_proof_m1]
      properties:
        email:
          type: string
          format: email
        client_public_key_a:
          type: string
          description: Base64 encoded SRP client public key A.
        client_proof_m1:
          type: string
          description: Base64 encoded SRP client proof M1.
        totp_code:
          type: string
          nullable: true
        backup_code:
          type: string
          nullable: true
    LoginVerifyResponse:
      type: object
      required:
        - server_proof_m2
        - access_token
        - refresh_token
        - token_type
        - expires_in_seconds
      properties:
        server_proof_m2:
          type: string
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          example: Bearer
        expires_in_seconds:
          type: integer
          format: int64
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
    AuthUserResponse:
      type: object
      required:
        - id
        - username
        - display_name
        - email
        - email_verified
        - two_factor_enabled
        - account_status
        - token_version
        - created_at
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        display_name:
          type: string
        email:
          type: string
          format: email
        avatar_url:
          type: string
          nullable: true
        email_verified:
          type: boolean
        two_factor_enabled:
          type: boolean
        account_status:
          type: string
        token_version:
          type: integer
          format: int32
        created_at:
          type: string
          format: date-time
    EmailVerificationRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
    EmailVerificationConfirmRequest:
      type: object
      required: [email, code]
      properties:
        email:
          type: string
          format: email
        code:
          type: string
          minLength: 6
          maxLength: 6
    StatusResponse:
      type: object
      required: [status]
      properties:
        status:
          type: string
    TwoFactorSetupResponse:
      type: object
      required: [secret, otpauth_url, backup_codes, expires_at]
      properties:
        secret:
          type: string
        otpauth_url:
          type: string
        backup_codes:
          type: array
          items:
            type: string
        expires_at:
          type: string
          format: date-time
    TwoFactorConfirmRequest:
      type: object
      required: [code]
      properties:
        code:
          type: string
          minLength: 6
          maxLength: 8
    TwoFactorDisableRequest:
      type: object
      properties:
        code:
          type: string
          nullable: true
        backup_code:
          type: string
          nullable: true
...